Added pidToPWM function

+++ ABS/abs.h
@@ -10,7 +10,7 @@ typedef enum { SPD_ON_TARGET,   // 0
                SPD_UNDER_TARGET,// 2
             } AbsEvent;
 
-AbsState absControl(float setpoint, float currentSpeedFilt, float* controlCmd, const PIDParams* params);   
+AbsState absControl(float setpoint, float currentSpeedFilt, uint8_t* controlCmd, const PIDParams* params);   
 
 /* ABS Parameters */
 #define ABS_WHL_SPEED_HYST 1.0f
